+++ ChangeLog.md
@@ -8,6 +8,8 @@
   information. ([#210](https://github.com/davep/rogallo/pull/210))
 - Added support for the `finger` protocol.
   ([#212](https://github.com/davep/rogallo/pull/212))
+- Added `!finger <user[@host]>` as a command line command.
+  ([#212](https://github.com/davep/rogallo/pull/212))
 
 ## v1.0.0